+++ kernel_tuner/__init__.py
@@ -1,3 +1,5 @@
 
+from kernel_tuner.integration import store_results
 
 from kernel_tuner.interface import tune_kernel, run_kernel
+